rename dataset to sr_dataset as timm installation includes huggingface datasets which cause problems
