+++ configs/data/vdr_dense_rerun.yaml
@@ -1,0 +1,9 @@
+!!python/object:options.Options
+dataset_path: data/vdr
+tuple_info_file_location: data_splits/vdr/
+dataset_scan_split_file: data_splits/vdr/scans.txt
+dataset: vdr
+mv_tuple_file_suffix: _eight_view_deepvmvs_dense.txt
+num_images_in_tuple: 8
+frame_tuple_type: dense
+split: test
