+++ sr_datasets/scanniverse_dataset.py
@@ -5,7 +5,7 @@ import numpy as np
 import PIL.Image as pil
 import torch
 from scipy.spatial.transform import Rotation as R
-from datasets.generic_mvs_dataset import GenericMVSDataset
+from sr_datasets.generic_mvs_dataset import GenericMVSDataset
 from torchvision import transforms
 from utils.geometry_utils import rotx
